+++ config/settings/local.py
@@ -71,6 +71,11 @@ DEBUG_TOOLBAR_CONFIG = {
 # ------------------------------------------------------------------------------
 INSTALLED_APPS += ['django_extensions', ]
 
+GRAPH_MODELS = {
+  'all_applications': False,
+  'group_models': True,
+}
+
 # TESTING
 # ------------------------------------------------------------------------------
 TEST_RUNNER = 'django.test.runner.DiscoverRunner'